refactor: split ingestion into src/ingestion submodules
Move run identity, aggregate, job worker, orchestration, and services into
src/ingestion/*.rs with src/ingestion.rs as the module root. No behavior change.
